HDF5File now loads the library on construction (#6)
This prevents errors where developers forgot to load the library manually.  It is still recommended to manually load the library in order to fail fast quickly and in a controlled way.

It was necessary to change HDF5File.OpenMode to keep a Supplier instead of an int, this is intended to delay classloading of H5 which will attempt and fail to load the native code on load.